+++ packages/desktop/app/views/ImagingRequestListingView.js
@@ -11,7 +11,6 @@ export const ImagingRequestListingView = React.memo(() => {
     <PageContainer>
       <TopBar title="Imaging requests" />
       <ImagingRequestsSearchBar
-        searchParameters={searchParameters}
         setSearchParameters={setSearchParameters}
       />
       <ImagingRequestsTable searchParameters={searchParameters} />